auth_service/requirements.auth.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
psycopg2-binary
redis
python-dotenv
pydantic-settings
pydantic[email]
PyJWT
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
pytest
pytest-asyncio
httpx
asgi_lifespan
aiosqlite
fastapi-pagination
jwcrypto==1.5.6
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-sqlalchemy
ruff
mypy